+++ source/MIBRR/src/MibrrData.cpp
@@ -1,12 +1,12 @@
 // Title:    Function definitions for the MibrrData Class
 // Author:   Kyle M. Lang
 // Created:  2014-AUG-24
-// Modified: 2017-NOV-30
+// Modified: 2018-FEB-12
 // Purpose:  This class contains the data-related functions used by the MIBRR
 //           Gibbs sampler.
 
 //--------------------- COPYRIGHT & LICENSING INFORMATION ---------------------//
-//  Copyright (C) 2017 Kyle M. Lang <k.m.lang@uvt.nl>                          //  
+//  Copyright (C) 2018 Kyle M. Lang <k.m.lang@uvt.nl>                          //  
 //                                                                             //
 //  This file is part of MIBRR.                                                //
 //                                                                             //
@@ -24,7 +24,7 @@
 //  with this program. If not, see <http://www.gnu.org/licenses/>.             //
 //-----------------------------------------------------------------------------//
 
-#include "MibrrData.hpp"
+#include "MibrrData.h"
 
 ///////////////////////// CONSTRUCTORS / DESTRUCTOR /////////////////////////////