+++ source/MIBRR/src/MibrrData.h
@@ -1,12 +1,12 @@
 // Title:    Header file for MibrrData Class
 // Author:   Kyle M. Lang
 // Created:  2014-AUG-24
-// Modified: 2017-NOV-27
+// Modified: 2018-FEB-12
 // Purpose:  This class contains data- and sampling-related functions used by the
 //           MIBRR Gibbs sampler.
 
 //--------------------- COPYRIGHT & LICENSING INFORMATION ---------------------//
-//  Copyright (C) 2017 Kyle M. Lang <k.m.lang@uvt.nl>                          //  
+//  Copyright (C) 2018 Kyle M. Lang <k.m.lang@uvt.nl>                          //  
 //                                                                             //
 //  This file is part of MIBRR.                                                //
 //                                                                             //
@@ -27,7 +27,7 @@
 #ifndef MIBRRDATA_H
 #define MIBRRDATA_H
 
-#include "MibrrDefs.hpp"
+#include "MibrrDefs.h"
 #include <algorithm>
 
 class MibrrData {