+++ source/MIBRR/src/runGibbs.cpp
@@ -1,12 +1,12 @@
 // Title:    Gibbs Sampler for MIBEN & MIBL
 // Author:   Kyle M. Lang
 // Created:  2014-AUG-20
-// Modified: 2017-NOV-27
+// Modified: 2018-FEB-12
 // Purpose:  This function will do the Gibbs sampling for the Bayesian Elastic
 //           Net and Bayesian LASSO models that underlie MIBRR's core functions.
 
 //--------------------- COPYRIGHT & LICENSING INFORMATION ---------------------//
-//  Copyright (C) 2017 Kyle M. Lang <k.m.lang@uvt.nl>                          //  
+//  Copyright (C) 2018 Kyle M. Lang <k.m.lang@uvt.nl>                          //  
 //                                                                             //
 //  This file is part of MIBRR.                                                //
 //                                                                             //
@@ -24,34 +24,34 @@
 //  with this program. If not, see <http://www.gnu.org/licenses/>.             //
 //-----------------------------------------------------------------------------//
 
-#include "MibrrData.hpp"
-#include "MibrrGibbs.hpp"
+#include "MibrrData.h"
+#include "MibrrGibbs.h"
 
 // The following plugin should allow us to use C++11:
 // [[Rcpp::plugins(cpp11)]]
 
 // [[Rcpp::export]]
-Rcpp::List runGibbs(Eigen::MatrixXd data,
-		    Eigen::VectorXd dataScales,
-		    int             nTargets,  
-		    Rcpp::List      missList,
-		    Eigen::VectorXi respCounts,
-		    Eigen::VectorXd lambda1,
-		    Eigen::VectorXd lambda2,
-		    Eigen::VectorXd l1Parms,
-		    Eigen::VectorXd l2Parms,
-		    Eigen::VectorXd sigmaStarts,
-		    Eigen::MatrixXd tauStarts,
-		    Eigen::MatrixXd betaStarts,
-		    int             burnSams,
-		    int             totalSams,
-		    bool            verbose,
-		    bool            doBl,
-		    bool            fullBayes,
-		    bool            adaptScales,
-		    bool            simpleIntercept,
-		    bool            noMiss,
-		    Eigen::VectorXi seeds)
+Rcpp::List runGibbs(Eigen::MatrixXd           data,
+		    Eigen::VectorXd           dataScales,
+		    int                       nTargets,  
+		    Rcpp::List                missList,
+		    Eigen::VectorXi           respCounts,
+		    Eigen::VectorXd           lambda1,
+		    Eigen::VectorXd           lambda2,
+		    Eigen::VectorXd           l1Parms,
+		    Eigen::VectorXd           l2Parms,
+		    Eigen::VectorXd           sigmaStarts,
+		    Eigen::MatrixXd           tauStarts,
+		    Eigen::MatrixXd           betaStarts,
+		    int                       burnSams,
+		    int                       totalSams,
+		    bool                      verbose,
+		    bool                      doBl,
+		    bool                      fullBayes,
+		    bool                      adaptScales,
+		    bool                      simpleIntercept,
+		    bool                      noMiss,
+		    std::vector<unsigned int> seeds)
 {
   // Unpack the list of missing row indices:
   std::vector< std::vector<int> > missIndices;